Migrate to GitHub Actions (#31)
This commit migrates the drone CI configuration to GitHub Actions.

Per the "docker/build-push-action" documentation, any changes made
to files preceding the build-push Action will be discarded. As a result,
any resolved git LFS files will be ignored. Because the repo. contains
one or more git LFS'ed tar file, the file(s) created by the "checkout"
action will be returned to LFS pointer text files when the "build-push"
action runs. This does not bode well for the Docker image build process,
which attempts to untar said pointer files. The documentation states
that this behavior is disabled when users override the Docker build
context, which will populate the resolved LFS file(s) into the build ctx.

Refer to the following URLs for more information:

- https://github.com/docker/build-push-action/#usage
- https://stackoverflow.com/a/66453723